pyproject.toml
[project]
name = "a1-compiler"
version = "0.1.5"
description = "An optimizing agent-to-code compiler"
readme = "README.md"
authors = [
    { name = "Caleb Winston", email = "calebhwin@gmail.com" }
]
requires-python = ">=3.12"
keywords = ["llm", "agents", "ai", "compiler", "code-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.10.0",
    "any-llm-sdk[all]>=1.0.0",
    "mcp-use>=0.1.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "fsspec>=2024.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
a1 = "a1.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/a1"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = ["-v", "--tb=short"]
markers = [
    "show_generated_code: Show generated code on test failure (automatically enabled with --show-generated-code flag)"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "duckdb>=1.0.0",
]

[dependency-groups]
dev = [
    "duckdb-engine>=0.17.0",
    "pre-commit>=4.4.0",
    "twine>=6.2.0",
]
